refactor(tests): split TranscriptionPostProcessor tests into domain-focused files
- break up `TranscriptionPostProcessorTests.swift` into focused extensions for:
  - email normalization/list interactions
  - capitalization and time normalization
  - language/laughter/idiom heuristics
- keep a slim core test file for shared/general post-processor behavior
- wire new test files into the `KeyVoxTests` target in the Xcode project
- preserve existing test coverage and behavior (focused TranscriptionPostProcessor suite passes: 63/63)
